fix(tailwind): tw-* variables in non inlined rules
