+++ packages/react-email/src/components/tailwind/utils/css/make-inline-styles-for.ts
@@ -1,6 +1,7 @@
 import { type CssNode, type Declaration, generate, walk } from 'css-tree';
 import { getReactProperty } from '../compatibility/get-react-property.js';
 import type { CustomProperties } from './get-custom-properties.js';
+import { stripEmptyTailwindVars } from './strip-empty-tailwind-vars.js';
 import { unwrapValue } from './unwrap-value.js';
 
 export function makeInlineStylesFor(
@@ -59,68 +60,7 @@ export function makeInlineStylesFor(
         if (declaration.property.startsWith('--')) {
           return;
         }
-        // Tailwind v4 emits variant-stacking idioms like
-        //   font-variant-numeric: var(--tw-ordinal,) var(--tw-slashed-zero,) tabular-nums var(--tw-numeric-fraction,)
-        // where each var() has an empty fallback so missing variants collapse to nothing.
-        // The walker above replaces var() calls with an initialValue when one is defined,
-        // but Tailwind deliberately leaves these variant vars undefined until used, so they
-        // stay in the output here and produce unresolvable custom properties in email HTML
-        // (no email client supports CSS custom properties reliably). Per the CSS spec
-        // (https://www.w3.org/TR/css-variables-1/#using-variables) an empty fallback means
-        // "use empty string if the variable is undefined", which is exactly what we want at
-        // inline-style time.
-        //
-        // Scoped to the `--tw-` prefix so any user-authored empty-fallback var() refs
-        // (even ones used inside tailwind utilities) are left untouched.
-        walk(declaration.value, {
-          visit: 'Function',
-          leave(func, funcItem, funcList) {
-            if (func.name !== 'var') {
-              return;
-            }
-
-            let variableName: string | undefined;
-            walk(func, {
-              visit: 'Identifier',
-              enter(identifier) {
-                variableName = identifier.name;
-                return this.break;
-              },
-            });
-            if (!variableName?.startsWith('--tw-')) {
-              return;
-            }
-
-            let sawComma = false;
-            let hasFallbackContent = false;
-            func.children.forEach((child) => {
-              if (!sawComma) {
-                if (child.type === 'Operator' && child.value === ',') {
-                  sawComma = true;
-                }
-                return;
-              }
-
-              let childValue = generate(child).trim();
-              if (child.type === 'Raw') {
-                const emptyTailwindVarPattern = /var\(--tw-[^,()]+,\s*\)/g;
-                childValue = childValue
-                  .replaceAll(emptyTailwindVarPattern, '')
-                  .trim();
-              }
-
-              if (childValue.length > 0) {
-                hasFallbackContent = true;
-              }
-            });
-
-            if (!sawComma || hasFallbackContent) {
-              return;
-            }
-
-            funcList.remove(funcItem);
-          },
-        });
+        stripEmptyTailwindVars(declaration.value);
 
         styles[getReactProperty(declaration.property)] =
           generate(declaration.value).trim() +
